requirements.txt
streamlit>=1.37.0
pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
nltk>=3.9
spacy>=3.7.5
transformers>=4.44.2
accelerate>=0.34.2
sentence-transformers>=3.0.1
bertopic>=0.16.3
umap-learn>=0.5.6
hdbscan>=0.8.38
plotly>=5.24.0
pydantic>=2.9.2
python-dotenv>=1.0.1
pyyaml>=6.0.2
xlrd>=2.0.1
openpyxl>=3.1.5
torch>=2.3.1
groq>=0.9.0
